refactor(TemplateDAO): Update to refactor TemplateDAO
