+++ bifrost/src/main/java/com/heimdallauth/server/service/EmailTemplateProcessor.java
@@ -1,6 +1,6 @@
 package com.heimdallauth.server.service;
 
-import com.heimdallauth.server.commons.models.TemplateModel;
+import com.heimdallauth.server.commons.models.bifrost.TemplateModel;
 import com.heimdallauth.server.datamanagers.TemplateDataManager;
 import com.heimdallauth.server.models.ProcessedEmailBodyModel;
 import lombok.extern.slf4j.Slf4j;
