+++ bifrost/src/main/java/com/heimdallauth/server/service/TemplateValidationService.java
@@ -1,13 +1,11 @@
 package com.heimdallauth.server.service;
 
-import com.heimdallauth.server.commons.models.TemplateModel;
+import com.heimdallauth.server.commons.models.bifrost.TemplateModel;
 import com.heimdallauth.server.datamanagers.TemplateDataManager;
 import com.heimdallauth.server.exceptions.TemplateAlreadyExists;
 import com.heimdallauth.server.exceptions.TemplateDoesNotExists;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
 
 @Service
 @Slf4j
